requirements.txt
pstats
snakeviz
torch_tb_profiler
rich
wandb